auth of mkemories created backup minute 50:52

+++ src/routes/momories.ts
@@ -1,10 +1,18 @@
 import { FastifyInstance } from 'fastify'
 import { prisma } from '../lib/prisma'
 import { z } from 'zod'
+import { request } from 'https'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async(request) => {
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -17,7 +25,8 @@ export async function memoriesRoutes(app: FastifyInstance) {
       }
     })
   })
-  app.get('/memories:id', async (request) => {
+
+  app.get('/memories:id', async (request, reply) => {
     const paramSchema = z.object({
       id: z.string().uuid(),
     })
@@ -29,6 +38,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     })
+
+    if (!memory.isPublic && memory.userId != request.user.sub){
+      return reply.status(401).send
+    }
 
     return memory
   })
@@ -46,25 +59,35 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '9013fbd0-3b8c-4af7-9296-fac14db7fb0f',
+        userId: request.user.sub,
       },
     })
     return memory
   })
-  app.put('/memories:id', async (request) => {
+
+  app.put('/memories:id', async (request, reply) => {
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
       isPublic: z.coerce.boolean().default(false),
     })
-    const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
-
     const paramSchema = z.object({
       id: z.string().uuid(),
     })
     const { id } = paramSchema.parse(request.params)
 
-    const memory = await prisma.memory.update({
+    const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
+
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+    if(memory.userId != request.user.sub){
+      return reply.status(401).send()
+    }
+    
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -77,11 +100,20 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories:id', async (request) => {
+  app.delete('/memories:id', async (request, reply) => {
     const paramSchema = z.object({
       id: z.string().uuid(),
     })
     const { id } = paramSchema.parse(request.params)
+    
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+    if (memory.userId != request.user.sub){
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
